fixed suncalculation & Remapping

+++ DVF-API/Data/Interfaces/ILocationRepository.cs
@@ -8,6 +8,6 @@ namespace DVF_API.Data.Interfaces
         Task<int> FetchLocationCount();
         Task<Dictionary<int, string>> FetchLocationCoordinates(int fromIndex, int toIndex);
         Task<List<BinaryDataFromFileDto>> FetchAddressByCoordinates(SearchDto searchDto);
-        Task<Dictionary<int, LocationDto>> GetAllLocationCoordinates();
+        Task<Dictionary<long, LocationDto>> GetAllLocationCoordinates();
     }
 }
